Rewrite of connection_result API
This API is now decoupled from the config_connection structure so that
it can be reused elsewhere (this will become necessary), and the writing
of single values is simplified to write directly into a given buffer.

This eliminates one inconsistency in the interface: in some places single
results are formatted via a callback, while in others a passed buffer is
written to.