Try to be consistent about using [] in function arguments
In a function definition x[] and *x are equivalent, but there is the
potential of some extra information.  Here I try to ensure that [] is
used when the argument is an array and * when it is a pointer to a single
value ... except for constant null terminated strings.  In this case it
seems to make more sense to treat the entire string as a single entity.