Another go at the the block interface
This time we're getting warmer, but it turns out that mapping memory pages
is not so simple, so at the moment table readbacks always return zero.
Also this version won't compile for simulation, but that'll be easy to fix.

Either we have to implement a vm page fault handler (good grief!) or
forget about memory mapping the memory.  Hmmm.