Fix warning reported by recent gcc
It looks more recent versions of gcc are back to flagging a -Wconversion
error on
    target <op>= value
when sizeof(target) < sizeof(int).  This is a known gcc bug,
    https://gcc.gnu.org/bugzilla/show_bug.cgi?id=40752
and it's *old* and so unlikely to get fixed permanently.

In this particular case we can easily work around it, and I'd like to hang
onto -Wconversion for now.