Fix nasty driver bug
It would seem that the method used to compute the register area base page
does not always work.  Running PandA on a smaller system (1GB instead of
2GB RAM) seems to be enough to trigger the issue: the address returned by
mapping /dev/panda.map does not map to the FPGA!

Fortunately the fix seems to be easy.

I so wish that there was clear and up to date documentation for this...