Fix automated counting of attributes array
Alas, the cross compiler we're using for PandA (gcc 4.9) doesn't properly
support the gcc initialiser array syntax in gnu99 mode.  This is a known
bug, the fix is already backported to RHEL7, see
    https://bugzilla.redhat.com/show_bug.cgi?id=1672652

Rather than changing the toolchain for this one application, we work around
it by counting the array length with a sentinel value instead.  This does
make the code a trifle more brittle, but that's the cost of this fix.