Change extension parse functions to return callables
This is a backwards incompatible change, but is very internal, so should be
acceptable.

Change the .parse_read and .parse_write methods to return callables
directly rather than going through an intermediate object with .read and
.write methods.  Makes a lot more sense.

At the same time, segregate the field numbers so that we don't mix up the
read and write methods.