Implement capture sharing for StdDev and Mean capture
This reduces data capture when both Mean and StdDev are required and
avoids capturing the same data twice.

Note that Sum is not shared, and thus if Sum and Mean or StdDev are
requested there will be duplication of data transfer.  This should not be
an issue, as this is an odd capture option!