Add support for driver compability check
A new register *DRV.DRV_COMPAT_VERSION has been added to the driver only
registers.  This reads as zero in current FPGA versions, but when an
incompatible change is made to the driver API this value will be updated.

The expected value is recorded in DRIVER_COMPAT_VERSION which is now
defined in the registers file.

This addresses in part issues #31 and PandABlocks/PandABlocks-FPGA#109