Fix GATE_DURATION duplication in RAW mode
`ensure_sample_count` is responsible for finding sample count if it
was already captured and we need it for a mean or std_dev capture.
The problem was that it was looking for it in the wrong capture group, it
should be the scaled32 group instead of the unscaled group.

This fixes PandABlocks/PandABlocks-server#47