Add option for some hardware registers to be optional
This commit allows some registers (specifically the two hardware timestamp
registers in this case) to be marked as optional, which means the server
will not complain if they are missing from the *REG (or *DRV) block.  This
works so long as the register can safely be read even if not implemented;
for the timestampl registers zero should be returned.

This allows the newer server to work with older hardware.