extension-server: notify systemd if needed
In the yocto build, panda-server and panda-extension-server are started
by systemd, in order to avoid wasteful loops trying to connect in the
server, this change makes the extension-server notify systemd that it is
ready to accept clients and then systemd can proceed to load the
panda-server following the service dependency specification.

This will not affect non-systemd workflow, as the notification is done
only if the variable NOTIFY_SOCKET (set by systemd) is present.