Fix polled attributes update index
It should be newer just compared to the current report index.

The reason why report_index + 1 is needed in this case is because it is
updated synchronously to getting the changes and we want the comparison
to be true for the current fetch changes operation but not for the next
one.