Fix polled attributes update index (#71)
* Fix polled attributes update index

It should be newer just compared to the current report index.

The reason why report_index + 1 is needed in this case is because it is
updated synchronously to getting the changes and we want the comparison
to be true for the current fetch changes operation but not for the next
one.

* Revert "tests: add bad test data to split the polled attribute change to a new PR"

This reverts commit 957ba01bbac97f0504822bcf8a4784d00e9f13c0.